feat: fix add expenses to user ID dynamically with user logged and refactor expenses and incomes pages with comprehensive UI components and improved user experience
