+++ src/app/expenses/components/add-expense-dialog.tsx
@@ -31,7 +31,7 @@ import { zodResolver } from "@hookform/resolvers/zod"
 import { useForm } from "react-hook-form"
 import * as z from "zod"
 import { useState, useEffect } from "react"
-import { Checkbox } from "../ui/checkbox"
+import { Checkbox } from "../../../components/ui/checkbox"
 import { createExpense } from "@/app/actions"
 import { NewExpense } from "@/types/expense"
 import { Loader2 } from "lucide-react"
