+++ src/app/expenses/components/edit-expense-dialog.tsx
@@ -30,7 +30,7 @@ import { zodResolver } from "@hookform/resolvers/zod"
 import { useForm } from "react-hook-form"
 import * as z from "zod"
 import { useState } from "react"
-import { Checkbox } from "../ui/checkbox"
+import { Checkbox } from "../../../components/ui/checkbox"
 import { updateExpense } from "@/app/actions"
 import { Expense } from "@/types/expense"
 import { Loader2, Pencil } from "lucide-react"
